Standardize redirect loop detection.
Standardize redirect loop detection so that all drivers throw a
``RedirectLoopException``, which knows the current url.
