+++ ftw/testbrowser/drivers/mechdriver.py
@@ -2,9 +2,11 @@ from ftw.testbrowser.drivers.utils import isolated
 from ftw.testbrowser.drivers.utils import remembering_for_reload
 from ftw.testbrowser.exceptions import BlankPage
 from ftw.testbrowser.exceptions import BrowserNotSetUpException
+from ftw.testbrowser.exceptions import RedirectLoopException
 from ftw.testbrowser.interfaces import IDriver
 from ftw.testbrowser.utils import copy_docs_from_interface
 from mechanize import Request
+from mechanize._urllib2_fork import HTTPRedirectHandler
 from requests.structures import CaseInsensitiveDict
 from zope.interface import implements
 import pkg_resources
@@ -60,6 +62,9 @@ class MechanizeDriver(object):
         try:
             self.response = self._get_mechbrowser().open(request)
         except urllib2.HTTPError as response:
+            if response.reason.startswith(HTTPRedirectHandler.inf_msg):
+                raise RedirectLoopException(response.geturl())
+
             self.response = response
         except:
             self.response = None
